Rework game tables, fix server-side bugs on check, add session support (crunch!)

+++ NiceNumber.Web/ViewModels/CheckModel.cs
@@ -5,7 +5,7 @@ namespace NiceNumber.Web.ViewModels
 {
     public class CheckModel
     {
-        public byte[] Positions { get; set; }
+        public int[] Positions { get; set; }
         
         public RegularityType Type { get; set; }
